preview canvas

+++ lib/app/models/text_model/text_font_model.dart
@@ -10,7 +10,7 @@ class TextFontModel {
   });
 
   String fontFamily;
-  int fontSize;
+  double fontSize;
   int fontWeight;
   // int color;
   String fontType;
